requirements.txt
typer==0.12.3
rich==13.7.1
python-dotenv==1.0.1
jinja2==3.1.4
pyyaml==6.0.2
openai==1.51.0
httpx==0.27.2
# Voice / GUI
vosk==0.3.45
pyaudio==0.2.14
Pillow==10.4.0
# For tkinter on most Linux distros, python3-tk is installed via apt

# Optional: Whisper backend if you switch VOICE_ENGINE=whisper later
# openai-whisper==20231117
